Crate cfg_test_utils

Test utilities that can be used across multiple pallets. Providing all sorts of mock implementations for traits the pallets need and which can be used for mock environments

Modules

mocks

Mocks of traits for usage in pallet-tests. Also does contain implementations for Substrate based traits for testing

Macros

impl_mock_accountant

Exposes a struct $name that implements the trait Accountant. The struct expects one generic parameter that implements the fungibles traits Inspect, Mutate and Transfer. Furthermore, there exists a struct GenesisConfig that implements trait GenesisBuild that can be used like any other GenesisConfig to initialize state in the TestExternalities.

impl_mock_registry

Constants

TEST_PALLET_ID

A pallet id for testing. Can be used for a single point of storage for tokens in testing for example